% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{Calculates snacks}
\usage{
snacks(grapes, apples, burritos)
}
\arguments{
\item{grapes}{mass of grapes in grams}

\item{apples}{number of apples in a bucket}

\item{burritos}{the number of burritos available to eat}
}
\description{
Calculates snacks
}
\examples{
snacks(grapes = 5, apples = 10, burritos = 2)
}
