% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thistle.R
\name{thistle}
\alias{thistle}
\title{The type and number of treats you are giving thistle}
\usage{
thistle(treat, number)
}
\arguments{
\item{treat}{the type of treat you are giving Thistle}

\item{number}{the number of treat you are giving Thistle}
}
\value{
none
}
\description{
The type and number of treats you are giving thistle
}
\examples{
thistle(treat = "tundra treat", number = 2)
}
